# contracts/openapi.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost","description":"Generated server url"}],"tags":[{"name":"Job Matching","description":"Endpoints for enriching found jobs with application state"}],"paths":{"/api/v1/job-matches/enrich":{"post":{"tags":["Job Matching"],"summary":"Enrich job search results","operationId":"enrichJobs","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EnrichJobsRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EnrichJobsResponse"}}}}}}}},"components":{"schemas":{"EnrichJobsRequest":{"type":"object","properties":{"userId":{"type":"string"},"jobs":{"type":"array","items":{"$ref":"#/components/schemas/JobMatchJob"}}}},"JobMatchJob":{"type":"object","properties":{"id":{"type":"string"},"canonicalJobId":{"type":"string"},"provider":{"type":"string"},"primarySource":{"type":"string"},"externalJobId":{"type":"string"},"title":{"type":"string"},"jobTitle":{"type":"string"},"company":{"type":"string"},"companyName":{"type":"string"},"location":{"type":"string"},"canonicalLocation":{"type":"object"},"salary":{"type":"object"},"employmentType":{"type":"string"},"postedDate":{"type":"string"},"description":{"type":"string"},"url":{"type":"string"},"sourceUrl":{"type":"string"},"sources":{"type":"array","items":{"$ref":"#/components/schemas/JobSourceReference"}},"distanceMiles":{"type":"number","format":"double"},"matchScore":{"type":"number","format":"double"},"applicationStatus":{"type":"string","enum":["NEW","DOCUMENTS_GENERATED","APPLIED","INTERVIEW","UNSUCCESSFUL","OFFER","ACCEPTED","REJECTED_BY_USER","WITHDRAWN"]},"applicationId":{"type":"string","format":"uuid"},"cvDocumentId":{"type":"string"},"coverLetterDocumentId":{"type":"string"},"appliedAt":{"type":"string","format":"date-time"},"applicationUpdatedAt":{"type":"string","format":"date-time"}}},"JobSourceReference":{"type":"object","properties":{"integrationProvider":{"type":"string"},"provider":{"type":"string"},"publisher":{"type":"string"},"externalJobId":{"type":"string"},"listingUrl":{"type":"string"},"applyUrl":{"type":"string"},"directApply":{"type":"boolean"},"providerPostedAt":{"type":"string","format":"date-time"}}},"EnrichJobsResponse":{"type":"object","properties":{"userId":{"type":"string"},"jobs":{"type":"array","items":{"$ref":"#/components/schemas/JobMatchJob"}}}}}}}